+++ tibiakt-core/src/main/kotlin/com/galarzaa/tibiakt/core/section/community/character/model/CharacterInfo.kt
@@ -17,9 +17,9 @@
 
 package com.galarzaa.tibiakt.core.section.community.character.model
 
-import com.galarzaa.tibiakt.core.domain.character.BaseCharacter
-import com.galarzaa.tibiakt.core.domain.character.CharacterLevel
+import com.galarzaa.tibiakt.core.domain.character.LevelAware
 import com.galarzaa.tibiakt.core.domain.character.Sex
+import com.galarzaa.tibiakt.core.domain.character.TibiaCharacter
 import com.galarzaa.tibiakt.core.domain.character.Vocation
 import com.galarzaa.tibiakt.core.section.community.urls.characterUrl
 import kotlinx.serialization.Serializable
@@ -54,7 +54,7 @@ import kotlin.time.Instant
  * @property otherCharacters The list of visible characters in the same account. Might be [isHidden].
  */
 @Serializable
-public data class Character(
+public data class CharacterInfo(
     override val name: String,
     val title: String?,
     val formerNames: List<String>,
@@ -80,7 +80,7 @@ public data class Character(
     val deaths: List<Death>,
     val accountInformation: AccountInformation?,
     val otherCharacters: List<OtherCharacter>,
-) : BaseCharacter, CharacterLevel {
+) : TibiaCharacter, LevelAware {
 
     /**
      * Whether this character is scheduled for deletion or nto.